refactored get and post to use ResponseDTO.
